Refactor
This commit moves quite a lot of the logic around and introduces some
new classes

Normalization now takes places inside an expression object, the
validation is far more isolated in single purpose validator objects.

CronParser knows the most as it needs to put together the ingredients
for Formatter to do it's thing, but both CronParser and Formatter know
less about things they don't need to, so that's good.